requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.8.1
attrs==25.3.0
billiard==4.2.1
bitarray==3.4.2
celery==5.3.6
certifi==2025.4.26
charset-normalizer==3.4.2
ckzg==2.1.1
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cytoolz==1.0.1
Django==4.2.11
django-cors-headers==4.3.1
django-debug-toolbar==4.2.0
django-filter==23.5
django-modelcluster==6.2.1
django-redis==5.4.0
django-taggit==4.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.5.0
drf-spectacular==0.27.1
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.0
eth_abi==5.2.0
frozenlist==1.7.0
gunicorn==21.2.0
hexbytes==1.3.1
idna==3.10
inflection==0.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kombu==5.5.4
multidict==6.4.4
packaging==25.0
parsimonious==0.10.0
pillow==10.2.0
prompt_toolkit==3.0.51
propcache==0.3.2
psycopg2-binary==2.9.9
pycryptodome==3.23.0
pydantic==2.11.5
pydantic_core==2.33.2
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2025.2
pyunormalize==16.0.0
PyYAML==6.0.2
redis==5.0.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
rlp==4.1.0
rpds-py==0.25.1
sentry-sdk==1.43.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
toolz==1.0.0
types-requests==2.32.4.20250611
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
web3==7.12.0
websockets==15.0.1
wheel==0.45.1
yarl==1.20.1
django-extensions==3.2.3
stripe==12.2.0
whitenoise==6.6.0
